Replace code. Update SecurityRequirementsParameterIsMissingError struct.
